requirements.txt
flask
flask-cors
cloudinary
python-dotenv
Flask-SQLAlchemy
pymysql
requests